Preparation for principled scoring.
git-svn-id: https://pet.opendfki.de/repos/pet/main@119 4200e16c-5112-0410-ac55-d7fb557a720a

+++ cheap/inputchart.h
@@ -133,7 +133,7 @@ public:
   
   input_token *add_token(const string &tokenstring);
   input_token *add_token(int id, int start, int end, class full_form ff,
-                         string orth, int p, const postags &pos,
+                         string orth, double p, const postags &pos,
                          const tPaths &paths,
                          bool synthesized = false)
   {
